@@ -1,58 +1,58 @@
 {
   "name": "gamebryo-plugin-management",
-  "version": "0.4.5",
+  "version": "0.5.2",
   "description": "Management for gamebryo plugins",
   "main": "./out/index.js",
   "repository": "",
   "scripts": {
     "webpack": "webpack --config webpack.config.js --progress --profile --color",
-    "webpack_async": "webpack --config webpack_async.config.js --entry ./node_modules/loot/async.js --output-filename async.js",
     "copy_esptk": "copyfiles -f ./node_modules/esptk/build/Release/esptk.node ./dist",
     "copy_loot": "copyfiles -f ./node_modules/loot/build/Release/node-loot.node ./dist && copyfiles -f ./node_modules/loot/loot_api/loot.dll ./dist && copyfiles -f ./src/loot_icon.png ./dist",
-    "copy_async": "npm run webpack_async && copyfiles -f ./node_modules/loot/bindings.js ./dist",
+    "copy_async": "copyfiles -f ./node_modules/loot/async.js ./dist && copyfiles -f ./node_modules/loot/bindings.js ./dist",
+    "fix_async_path": "node -e \"const fs=require('fs');const f='dist/async.js';let c=fs.readFileSync(f,'utf8');c=c.replace('./build/Release/node-loot','./node-loot');fs.writeFileSync(f,c);\"",
     "copy_assets": "copyfiles -f src/stylesheets/plugin_management.scss dist && copyfiles -f src/language.json dist",
-    "build": "npm run webpack && npm run copy_esptk && npm run copy_loot && npm run copy_async && npm run copy_assets && extractInfo",
+    "build": "npm run webpack && npm run copy_esptk && npm run copy_loot && npm run copy_async && npm run fix_async_path && npm run copy_assets && extractInfo",
     "installDev": "copyfiles -u 1 dist/**/* ../../out/bundledPlugins/gamebryo-plugin-management",
     "buildDev": "npm run build && npm run installDev"
   },
   "author": "Black Tree Gaming Ltd.",
   "license": "GPL-3.0",
   "devDependencies": {
-    "@types/cytoscape": "3.2.7",
-    "@types/immutability-helper": "^2.6.3",
-    "@types/js-yaml": "^3.11.2",
-    "@types/react": "^16.4.11",
+    "@types/cytoscape": "^3.19.16",
+    "@types/js-yaml": "^4.0.9",
+    "@types/react": "^16.14.34",
     "@types/react-bootstrap": "^0.32.25",
-    "@types/react-dom": "^16.0.7",
+    "@types/react-dom": "^16.9.24",
     "@types/react-faux-dom": "^4.1.1",
-    "@types/react-redux": "^6.0.6",
-    "@types/react-router": "^4.0.30",
-    "@types/react-select": "^1.2.1",
+    "@types/react-redux": "^7.1.33",
+    "@types/react-router": "^5.1.20",
+    "@types/react-select": "^3.0.26",
     "@types/redux": "^3.6.0",
     "@types/redux-thunk": "^2.1.0",
-    "bluebird": "^3.5.1",
-    "copyfiles": "^2.1.1",
-    "cytoscape": "^3.2.16",
-    "cytoscape-cose-bilkent": "^4.0.0",
+    "bluebird": "^3.7.2",
+    "copyfiles": "^2.4.1",
+    "cytoscape": "^3.6.0",
+    "cytoscape-cose-bilkent": "^4.1.0",
     "cytoscape-edgehandles": "^3.5.1",
-    "esptk": "Nexus-Mods/node-esptk#15918-starfield-update",
+    "esptk": "Nexus-Mods/node-esptk",
     "i18next": "^19.8.4",
-    "immutability-helper": "^2.7.1",
-    "js-yaml": "^3.12.0",
+    "immutability-helper": "^3.1.1",
+    "js-yaml": "^4.1.0",
     "json-loader": "^0.5.7",
     "loot": "Nexus-Mods/node-loot",
     "markdown-ast": "^0.2.1",
-    "prop-types": "^15.6.2",
-    "react": "^16.8.6",
-    "react-dnd": "^7.4.5",
-    "react-dom": "^16.8.6",
-    "react-i18next": "^11.8.5",
-    "react-select": "^1.2.1",
-    "redux": "^4.0.0",
-    "redux-thunk": "^2.3.0",
-    "ts-loader": "^9.2.6",
+    "prop-types": "^15.8.1",
+    "react": "^16.14.0",
+    "react-dnd": "^14.0.5",
+    "react-dom": "^16.14.0",
+    "react-i18next": "^15.0.2",
+    "react-select": "^5.8.1",
+    "redux": "^4.2.1",
+    "redux-act": "^1.8.0",
+    "redux-thunk": "^2.4.2",
+    "ts-loader": "^9.5.1",
     "vortex-api": "Nexus-Mods/vortex-api",
-    "webpack": "^5.76.0",
-    "webpack-cli": "^4.9.1"
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4"
   }
 }
